refactor(renderer): gather the sidebar domain into features/sidebar
The sidebar was six places at once: the domain context at the renderer root,
four pure-logic modules beside it, the prop-assembling pane, the entry component
in `components/`, and a 26-file part directory with its own `stores/`. Two
`stores/` directories for one domain, in two layers of the tree. Reading "fix
the pin ordering" meant finding all six.

This is the first `features/<domain>/` directory: the context, the logic, the
view, the parts and the tests in one place, flat except for `stores/`, which is
the shape `components/sidebar/` and `components/conversation/` already had.
No barrel, deep relative imports, tests co-located - all unchanged.

`sidebar.tsx` becomes `sidebar-context.tsx`. It has to: `sidebar.tsx` and
`Sidebar.tsx` coexisted only because they were in different directories, and
collapsing them into one collides on case-insensitive APFS. That would have
failed the macOS `Check` job alone and looked unrelated to the move.

Pure move - no behaviour, no assertion and no test count changed. 37 sidebar
tests and 51 App tests pass; `typecheck:renderer` covers every rewritten import,
since the renderer has no path aliases.
