+++ src/renderer/src/app-providers.tsx
@@ -14,6 +14,7 @@ import { ConversationProvider } from "./conversation";
 import { DirectMessagesProvider } from "./direct-messages";
 import { DynamicIslandProvider } from "./dynamic-island";
 import { DynamicIslandBridge } from "./dynamic-island-bridge";
+import { SidebarProvider } from "./features/sidebar/sidebar-context";
 import { LayoutProvider } from "./layout";
 import { NavigationProvider } from "./navigation";
 import { PlatformProvider } from "./platform";
@@ -27,7 +28,6 @@ import { ServerSwitchProvider } from "./server-switch";
 import { ServersProvider, useServers } from "./servers";
 import { SettingsProvider } from "./settings";
 import { SetupProvider } from "./setup";
-import { SidebarProvider } from "./sidebar";
 import { notifyTeamTyping } from "./team-typing";
 import { TurnsProvider } from "./turns";
 import { UiErrorsProvider } from "./ui-errors";
