+++ src/renderer/src/features/sidebar/SidebarAgentSection.tsx
