+++ src/renderer/src/features/sidebar/SidebarIcons.tsx
