+++ src/renderer/src/features/sidebar/SidebarPeopleSection.tsx
@@ -5,8 +5,8 @@
  */
 
 import { For, Show } from "solid-js";
-import { TeamPersonAvatar, teamMemberName } from "../TeamPersonAvatar";
-import { Badge, Button } from "../ui";
+import { TeamPersonAvatar, teamMemberName } from "../../components/TeamPersonAvatar";
+import { Badge, Button } from "../../components/ui";
 import { SidebarSectionHeader } from "./SidebarSectionHeader";
 import { sidebarMessageTime } from "./sidebar-filtering";
 import { useSidebarScope } from "./sidebar-scope";
