+++ src/renderer/src/features/sidebar/SidebarSectionHeader.tsx
@@ -5,7 +5,7 @@
  */
 
 import { Show } from "solid-js";
-import { ArrowDown, ArrowUp, buttonVariants, ChevronDown, ContextMenu, Pencil, Trash2 } from "../ui";
+import { ArrowDown, ArrowUp, buttonVariants, ChevronDown, ContextMenu, Pencil, Trash2 } from "../../components/ui";
 import { SidebarSectionEditor } from "./SidebarSectionEditor";
 import { useSidebarScope } from "./sidebar-scope";
