+++ src/renderer/src/features/sidebar/SidebarSectionList.tsx
