+++ src/renderer/src/features/sidebar/sidebar-agent-states.test.ts
