+++ src/renderer/src/features/sidebar/sidebar-drag-dom.ts
