+++ src/renderer/src/features/sidebar/sidebar-drag-model.ts
