+++ src/renderer/src/features/sidebar/sidebar-pins.test.ts
