+++ src/renderer/src/features/sidebar/sidebar-pins.ts
