+++ src/renderer/src/features/sidebar/sidebar-sections.test.ts
