+++ src/renderer/src/features/sidebar/sidebar-sections.ts
