+++ src/renderer/src/features/sidebar/stores/search-store.ts
@@ -4,7 +4,7 @@
  */
 
 import { createMemo, createSignal } from "solid-js";
-import type { SidebarProps } from "../../Sidebar";
+import type { SidebarProps } from "../Sidebar";
 
 export function createSidebarSearchStore(deps: { props: SidebarProps }) {
   const [query, setQuery] = createSignal("");
